Finished refactoring

+++ server/src/main/java/sw806f18/server/api/SurveyResource.java
@@ -5,6 +5,7 @@ import org.springframework.core.io.InputStreamResource;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import sw806f18.server.Authentication;
 import sw806f18.server.Constants;
 import sw806f18.server.database.Database;
 import sw806f18.server.database.NoSqlDatabase;
@@ -15,6 +16,7 @@ import javax.websocket.server.PathParam;
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import java.io.UnsupportedEncodingException;
+import java.lang.Error;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
 
@@ -39,8 +41,8 @@ public class SurveyResource {
         Survey survey = Database.getSurvey(id);
 
         InputStreamResource inputStreamResource =
-            new InputStreamResource(new ByteArrayInputStream(
-                survey.getHTML().getBytes(StandardCharsets.UTF_8)));
+                new InputStreamResource(new ByteArrayInputStream(
+                        survey.getHTML().getBytes(StandardCharsets.UTF_8)));
         return ResponseEntity.ok(inputStreamResource);
     }
 
@@ -53,15 +55,15 @@ public class SurveyResource {
      * @return stream
      */
     @RequestMapping(path = "/{id}", method = RequestMethod.POST,
-        consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE,
-        produces = MediaType.TEXT_HTML_VALUE)
+            consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE,
+            produces = MediaType.TEXT_HTML_VALUE)
     public ResponseEntity postSurvey(@PathVariable("id") int id, @RequestBody String params)
-        throws UnsupportedEncodingException {
+            throws UnsupportedEncodingException {
         InputStream stream;
         Survey survey = Database.getSurvey(id);
         boolean hasWarnings = false;
 
-        String content = new String(Base64.getDecoder().decode(params.substring(0, params.length() - 2)));
+        String content = new String(Base64.getDecoder().decode(params.substring(0, params.indexOf("="))));
 
         Map<String, String> formParams = new HashMap<>();
         String[] formMembers = content.split("&");
@@ -140,9 +142,9 @@ public class SurveyResource {
         builder.append("<script type='text/javascript'>");
 
         builder.append("window.onload = function() {\n"
-            + "    // similar behavior as clicking on a link\n"
-            + "    window.location.href = \"" + url + "\";\n"
-            + "}");
+                + "    // similar behavior as clicking on a link\n"
+                + "    window.location.href = \"" + url + "\";\n"
+                + "}");
 
         builder.append("</script>");
 
@@ -166,11 +168,11 @@ public class SurveyResource {
     public void postTestSurvey() {
 
         Survey survey1 = new Survey("Spørgeskema under graviditetsforløb",
-            "Dette spørgeskema indeholder spørgsmål vedrørende din livsstil og dit helbred.");
+                "Dette spørgeskema indeholder spørgsmål vedrørende din livsstil og dit helbred.");
         survey1.addStyleProperty("body", "background-color", "lightblue");
 
         survey1.addQuestion(new NumberQuestion(3, "Alkohol",
-            "Hvor mange genstande drikker du om ugen?"));
+                "Hvor mange genstande drikker du om ugen?"));
 
         List<String> bristolStoolChart = new ArrayList<>();
         bristolStoolChart.add("");
@@ -183,11 +185,11 @@ public class SurveyResource {
         bristolStoolChart.add("Type 7: Vandet og uden substans, fuldstændig flydende");
 
         survey1.addQuestion(new DropdownQuestion(2, Question.Type.STRING,
-            "Afføring",
-            "Hvordan vil du beskrive din afføring efter et gennemsnitligt toiletbesøg?",
-            bristolStoolChart));
+                "Afføring",
+                "Hvordan vil du beskrive din afføring efter et gennemsnitligt toiletbesøg?",
+                bristolStoolChart));
         survey1.addQuestion(new TextQuestion(1, "Sygdomsepisoder",
-            "Hvilke sygdomsepisoder har du haft inden for det sidste år?"));
+                "Hvilke sygdomsepisoder har du haft inden for det sidste år?"));
 
         List<String> rygningList = new ArrayList<>();
         rygningList.add("Dagligt mere end 4");
@@ -197,20 +199,20 @@ public class SurveyResource {
         rygningList.add("Aldrig");
 
         survey1.addQuestion(new DropdownQuestion(2, Question.Type.STRING,
-            "Rygning",
-            "Hvor ofte ryger du?",
-            rygningList));
+                "Rygning",
+                "Hvor ofte ryger du?",
+                rygningList));
 
         survey1.addQuestion(new TextQuestion("Graviditetsrelaterede begivenheder",
-            "Har du haft nogen specielle episoder?"));
+                "Har du haft nogen specielle episoder?"));
         survey1.addQuestion(new TextQuestion("Fødselsrelaterede begivenheder",
-            "Skete der noget specielt under fødslen?"));
+                "Skete der noget specielt under fødslen?"));
         survey1.addQuestion(new TextQuestion("Barselsrelaterede begivenheder",
-            "Er der sket noget i din tid der hjemme?"));
+                "Er der sket noget i din tid der hjemme?"));
 
 
         Survey survey2 = new Survey("Barn 6 - 10 år",
-            "Dette spørgeskema vedrører dit barn.");
+                "Dette spørgeskema vedrører dit barn.");
         survey2.addStyleProperty("body", "background-image", "url('https://media.istockphoto.com/photos/textured-blue-painted-background-picture-id534129318?k=6&m=534129318&s=612x612&w=0&h=5N2BeInhaXkV_G09cVoIaO2RWoNwGABqVbhw0U_0Jto=')");
         survey2.addStyleProperty("body", "background-size", "cover");
 
@@ -226,27 +228,27 @@ public class SurveyResource {
         ageChart.add("10 År");
 
         survey2.addQuestion(new DropdownQuestion(2, Question.Type.STRING,
-            "Barnets alder",
-            "Hvad er barnets alder?",
-            ageChart));
+                "Barnets alder",
+                "Hvad er barnets alder?",
+                ageChart));
         survey2.addQuestion(new NumberQuestion(3, "Barnets højde",
-            "Skriv barnets højde i centimeter:"));
+                "Skriv barnets højde i centimeter:"));
         survey2.addQuestion(new NumberQuestion(3, "Barnets vægt",
-            "Skriv barents vægt i gram:"));
+                "Skriv barents vægt i gram:"));
 
 
         Survey survey3 = new Survey("Efter fødsel til mor",
-            "Yderlige information om dig efter fødslen");
+                "Yderlige information om dig efter fødslen");
         survey3.addStyleProperty("body", "background-color", "#FAD7A0");
         survey3.addQuestion(new DropdownQuestion(2, Question.Type.STRING,
-            "Afføring",
-            "Hvordan vil du beskrive din afføring efter et gennemsnitligt toiletbesøg?",
-            bristolStoolChart));
+                "Afføring",
+                "Hvordan vil du beskrive din afføring efter et gennemsnitligt toiletbesøg?",
+                bristolStoolChart));
         survey3.addQuestion(new NumberQuestion(3, "Alkohol",
-            "Hvor mange genstande drikker du om ugen?"));
+                "Hvor mange genstande drikker du om ugen?"));
         survey3.addQuestion(new TextQuestion("Kost:",
-            "Hvad består din daglige kost af?"
-                + "\n\"Eks: Får du mange fibre? Spiser du sundt?\""));
+                "Hvad består din daglige kost af?"
+                        + "\n\"Eks: Får du mange fibre? Spiser du sundt?\""));
         survey3.addQuestion(new TextQuestion("Medicin", "Får du medicin?:"));
         survey3.addQuestion(new NumberQuestion("BMI", "Hvad er din BMI?:"));
 
@@ -265,13 +267,10 @@ public class SurveyResource {
      * @param id ID of survey to get.
      * @return Return Survey with ID.
      */
-    @GET
-    @Path("/{id}/object")
-    @Produces(MediaType.APPLICATION_JSON)
-    public Survey getSurveyObject(@PathParam("id") int id) {
+    @RequestMapping(method = RequestMethod.GET, path = "/{id}/object", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Survey> getSurveyObject(@PathParam("id") int id) {
         Survey survey = Database.getSurvey(id);
-
-        return survey;
+        return ResponseEntity.ok(survey);
     }
 
     /**
@@ -281,21 +280,22 @@ public class SurveyResource {
      * @param questionId Question to remove.
      * @return Response.
      */
-    @DELETE
-    @Path("/{surveyId}/question/{questionId}")
-    public Response deleteQuestionFromSurvey(@PathParam("surveyId") int surveyId,
-                                             @PathParam("questionId") int questionId) {
-        try {
-            Database.removeQuestionFromSurvey(surveyId, questionId);
+    @RequestMapping(method = RequestMethod.DELETE, path = "/{surveyId}/question/{questionId}",
+            produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity deleteQuestionFromSurvey(@PathVariable(value = "surveyId") int surveyId,
+                                                   @PathVariable(value = "questionId") int questionId,
+                                                   @CookieValue(value = "token") String token) {
+        if (Database.isResearcher(Authentication.instance.getId(token))) {
+            try {
+                Database.removeQuestionFromSurvey(surveyId, questionId);
 
-            Response response = Response.ok().build();
-            return response;
-        } catch (SurveyException e) {
-            e.printStackTrace();
-
-            Response response = Response.serverError().build();
-            return response;
+                return ResponseEntity.ok().build();
+            } catch (SurveyException e) {
+                e.printStackTrace();
+                return ResponseEntity.badRequest().body(new Error("Server Error"));
+            }
         }
+        return ResponseEntity.badRequest().body(new Error("Invalid token"));
     }
 }
